plotter works, keeps all things between restarts

+++ spansh_route.h
@@ -8,9 +8,9 @@ class SpanshRoute
 private:
     RestClient::parameters p;
 public:
-    SpanshRoute(float eff, float range, const std::string& from, const std::string& to)
+    SpanshRoute(uint32_t eff, float range, const std::string& from, const std::string& to)
     {
-        p["efficiency"] = stringfmt("%02f", eff);
+        p["efficiency"] = stringfmt("%u", eff);
         p["range"] = stringfmt("%02f", range);
         p["from"] = from;
         p["to"] = to;
